+++ CHANGELOG.md
@@ -39,5 +39,6 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ### Improvements
 
+* [#7](https://github.com/evmos/go-ethereum/pull/7) Implement custom active precompiles for the EVM.
 * [#3](https://github.com/evmos/go-ethereum/pull/3) Move the `JumpTable` defaults to a separate function.
 * [#2](https://github.com/evmos/go-ethereum/pull/2) Define `Interpreter` interface for the EVM.